+++ src/store/global.store.ts
